Move iteration-specific components, per guiding principles
