+++ client/my-sites/plans/jetpack-plans/spp/products-grid-spp/index.tsx
@@ -10,15 +10,22 @@ import { useSelector } from 'react-redux';
 /**
  * Internal dependencies
  */
-import { SWITCH_PLAN_SIDES_EXPERIMENT, SWITCH_PLAN_SIDES_TREATMENT } from '../experiments';
-import PlansFilterBarI5 from '../plans-filter-bar-i5';
-import ProductCardI5 from '../product-card-i5';
-import { getProductPosition } from '../product-grid/products-order';
-import { getPlansToDisplay, getProductsToDisplay, isConnectionFlow } from '../product-grid/utils';
-import useGetPlansGridProducts from '../use-get-plans-grid-products';
+import {
+	SWITCH_PLAN_SIDES_EXPERIMENT,
+	SWITCH_PLAN_SIDES_TREATMENT,
+} from 'calypso/my-sites/plans/jetpack-plans/experiments';
+import PlansFilterBarI5 from 'calypso/my-sites/plans/jetpack-plans/plans-filter-bar-i5';
+import ProductCardI5 from 'calypso/my-sites/plans/jetpack-plans/product-card-i5';
+import { getProductPosition } from 'calypso/my-sites/plans/jetpack-plans/product-grid/products-order';
+import {
+	getPlansToDisplay,
+	getProductsToDisplay,
+	isConnectionFlow,
+} from 'calypso/my-sites/plans/jetpack-plans/product-grid/utils';
+import useGetPlansGridProducts from 'calypso/my-sites/plans/jetpack-plans/use-get-plans-grid-products';
 import Experiment from 'calypso/components/experiment';
-import JetpackFreeCard from 'calypso/components/jetpack/card/jetpack-free-card-spp';
-import JetpackCrmFreeCardSPP from 'calypso/components/jetpack/card/jetpack-crm-free-card-spp';
+import JetpackFreeCard from 'calypso/components/jetpack/card/spp/jetpack-free-card-spp';
+import JetpackCrmFreeCardSPP from 'calypso/components/jetpack/card/spp/jetpack-crm-free-card-spp';
 import isJetpackCloud from 'calypso/lib/jetpack/is-jetpack-cloud';
 import {
 	PLAN_JETPACK_SECURITY_DAILY,
@@ -28,14 +35,17 @@ import { getCurrentUserCurrencyCode } from 'calypso/state/current-user/selectors
 import { getVariationForUser } from 'calypso/state/experiments/selectors';
 import getSitePlan from 'calypso/state/sites/selectors/get-site-plan';
 import getSelectedSiteId from 'calypso/state/ui/selectors/get-selected-site-id';
-import MoreInfoBox from '../more-info-box';
+import MoreInfoBox from 'calypso/my-sites/plans/jetpack-plans/more-info-box';
 import StoreFooter from 'calypso/jetpack-connect/store-footer';
 import getSiteId from 'calypso/state/selectors/get-site-id';
 
 /**
  * Type dependencies
  */
-import type { ProductsGridProps, SelectorProduct } from '../types';
+import type {
+	ProductsGridProps,
+	SelectorProduct,
+} from 'calypso/my-sites/plans/jetpack-plans/types';
 import type { JetpackProductSlug } from 'calypso/lib/products-values/types';
 import type { JetpackPlanSlugs } from 'calypso/lib/plans/types';
